Rename soroban-auth test names to describe tests
