+++ soroban-auth/src/tests/test_ed25519.rs
