+++ soroban-auth/src/tests/test_ed25519_with_nonce.rs
